+++ api/src/controllers/deleteAddressUser.js
@@ -2,28 +2,23 @@
    Author: Alejandro Téllez Aguilar
    Description: Crea la ruta /address/id para eliminar un domicilio del Usuari
 */
-
 const { Address } = require("../db");
+const {response} = require("../utils");
 
-const deleteAddressUser = async (id,userEmail) => {
+module.exports = async (req,res,next) => {
+    const {id, userEmail} = req.query;
     try {
-
         const street = await Address.findOne({
             where: {
                 id: id,
                 userEmail: userEmail
-                
             },
-          });
-          
-          if (street) {
+            });
+            if (street) {
             await street.destroy(); // deletes the row
-          }
-
-        return `Se elimino el domicilio con id: ${id} y usuario: ${userEmail}`;
+            }
+        return response(res,200,`Se elimino el domicilio con id: ${id} y usuario: ${userEmail}`);
     } catch (error) {
-        console.log(error);
-    }
-}
-
-module.exports = { deleteAddressUser };
+        next(error);
+    };
+};
