+++ api/src/controllers/getOrderStatus.js
@@ -2,21 +2,28 @@
    Author: Alejandro Téllez Aguilar
    Description: Crea el servicio de la ruta /addresses/?status/products para obtener las ordenes y sus productos
 */
-
 const { Order, Orderdetail } = require("../db");
+const {response} = require("../utils");
 
-const getOrderStatus = async (req, res, next) => {
-    const status = req.params.status
+module.exports = async (req, res, next) => {
+    const status = req.params.status;
     try {
-        const orderDb = await Order.findAll(status);
-        if (orderDb) {
-            Order.getHeadorder();
-
-        }
-        return res.json(userDb);
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-module.exports = { getOrderStatus };
+        const data = await Order.findOne({
+          where: { 
+              status: status,
+          },
+          include: [{//esta llave es por si tiene más relaciones, puede quitarse si sólo es una relación como en este caso
+              model: Orderdetail,
+              as: "headorder",
+              atributes:["purchasedamount", "purchaseprice"]
+          }]
+        }); 
+        if (data) {
+          return response(res,200,data);
+        } else {
+          return response(res,404,{message:"No se encontraron ordenes con ese status"});
+        }  
+      } catch (error) {
+          next(error);
+      };
+};
