+++ api/src/controllers/getProductById.js
@@ -1,12 +1,18 @@
 const { Product } = require("../db");
+const {response} = require('../utils')
 
-module.exports = async (id) => {
-    let detail = await Product.findByPk(id);
-        if (detail) {
-            return detail;
-        }
-        return {
-            message: 'Producto no Encontrado',
-            error: `El id: ${id} no existe`
-        };
+module.exports = async (req,res,next) => {
+    const {id} = req.params;
+    try {
+        const detail = await Product.findByPk(id);
+            if (detail) {
+                return response(res,200,detail);
+            }
+            return response(res,200,{
+                message: 'Producto no Encontrado',
+                error: `El id: ${id} no existe`
+            });
+    } catch (error) {
+      next(error);  
+    };
 };
