+++ api/src/controllers/getUserIdOrders.js
@@ -2,14 +2,28 @@
    Author: Alejandro Téllez Aguilar
    Description: Crea el servicio de la ruta /users/?id/orders para otener todos los usuarios y sus ordenes
 */
+const {User, Order} = require('../db');
+const {response} = require("../utils")
 
-const {User} = require("../db");
-
-module.exports = async (req, res) => {
+module.exports = async (req, res, next) => {
     const id = req.params.id;
-        const userDb = await User.findByPk(id);
-        if (userDb) {
-            User.getOrdenes();
-        }
-        return res.json(userDb);
+    try {
+        const data = await User.findOne({
+          where: { 
+              id: id,
+          },
+          include: [{//esta llave es por si tiene más relaciones, puede quitarse si sólo es una relación como en este caso
+              model: Order,
+              as: "ordenes",
+              atributes:["purchasecost", "payorder","paymentmethod"]
+          }]
+        }); 
+        if (data) {
+          return response(res,200,data);
+        } else {
+          return response(res,404,{message:"No se encontró usuario"});
+        }  
+      } catch (error) {
+          next(error);
+      };
 };
