+++ api/src/controllers/updateProduct.js
@@ -1,13 +1,24 @@
 const { Product } = require("../db");
+const {response} = require("../utils");
 
-const updateProduct = async (id) => {
+module.exports = async (req,res,next) => {
+    const {id} = req.params;
+    const { title, price, category, description, image, material, stock} = req.body;
     try {
-        let detail = await Product.findByPk(id);
-        console.log(detail);
-        return detail;
+        let update = await Product.findByPk(id);
+        update.title        = title;
+        update.price        = price;
+        update.category     = category;
+        update.description  = description;
+        update.image        = image;
+        update.material     = material;
+        update.stock        = stock;
+        console.log(update)
+        await update.save()
+        response(res,200,{
+            message: 'Actualizacion de Producto exitosa',
+        });
     } catch (error) {
-        console.log(error);
-    }
-}
-
-module.exports = { updateProduct };
+        next(error);
+    };
+};
