+++ api/src/controllers/updateUser.js
@@ -1,14 +1,20 @@
 const { User } = require("../db");
+const {response} = require("../utils");
 
-const updateUser = async (useremail) => {
+module.exports = async (req,res,next) => {
+    const {useremail} = req.params;
+    const { rol} = req.body;
     try {
-        let detail = await User.findByPk(useremail);
-        console.log(detail);
-        
-        return detail;
+        const update = await User.findByPk(useremail);
+        update.rol = rol;
+        await update.save()
+        response(res,200,{
+            message: 'Actualizacion de Usuario exitosa',
+            usuario: update.name,
+            nuevo_rol: update.rol,
+        });
     } catch (error) {
         console.log(error);
-    }
-}
-
-module.exports = { updateUser };
+        next(error);
+    };
+};
