+++ api/src/routes/deleteProduct.js
@@ -1,18 +1,8 @@
 const { Router } = require('express');
-const { deleteProduct } = require('../controllers/deleteProduct');
+const controller = require('../controllers');
 
 const router = Router();
 
-router.delete('/:id', async (req, res, next) => {
-    let {id} = req.params;
-    try {
-        let deleteP = await deleteProduct(id);
-        res.json({
-            'Eliminado': `${deleteP}`
-        });
-    } catch (error) {
-        next(error);
-    }
-})
+router.delete('/:id', controller.deleteProduct);
 
 module.exports = router;
