+++ api/src/routes/getOrderStatus.js
@@ -1,6 +1,5 @@
 const { Router } = require('express');
-const { getOrderStatus } = require('../controllers/getOrderStatus');
-const {Order, Orderdetail} = require('../db');
+const controller = require('../controllers');
 
 /* Date Creation: January 3, 2023
    Author: Alejandro Téllez Aguilar
@@ -8,29 +7,6 @@ const {Order, Orderdetail} = require('../db');
 */
 const router = Router();
 
-router.get('/:status/products', async (req, res, next) => {
-    const status = req.params.status;
-    try {
-      const data = await Order.findOne({
-        where: { 
-            status: status,
-        },
-        include: [{//esta llave es por si tiene más relaciones, puede quitarse si sólo es una relación como en este caso
-            model: Orderdetail,
-            as: "headorder",
-            atributes:["purchasedamount", "purchaseprice"]
-        }]
-        
-      }); 
-      if (data) {
-        return res.status(200).json(data);
-      } else {
-        return res.status(404).json({message:"No se encontraron ordenes con ese status"});
-      }  
-      
-    } catch (error) {
-        return res.status(404).json({error: error.message});
-    }
-});
+router.get('/:status/products', controller.getOrderStatus);
 
 module.exports = router;
