+++ api/src/routes/getProductById.js
@@ -1,13 +1,8 @@
 const { Router } = require('express');
 const controller = require('../controllers/index');
-const {response} = require('../utils')
 
 const router = Router();
 
-router.get('/:id', async (req, res) => {
-    let { id } = req.params;
-        let data = await controller.getProductById(id)
-        response(res,200,data);
-});
+router.get('/:id', controller.getProductById);
 
 module.exports = router;
