+++ api/src/utils/response.js
@@ -1,3 +1,6 @@
 module.exports = (res,statusCode,data)=> {
-    res.status(statusCode).json(data);
+    res.status(statusCode).json({
+        error: false,
+        data
+    });
 };
